+++ includes/Preview.php
@@ -2,10 +2,10 @@
 /**
  * Preview Class
  *
- * @package HeadlessCompanion
+ * @package Decouplix
  */
 
-namespace HeadlessCompanion;
+namespace Decouplix;
 
 defined( 'ABSPATH' ) || exit;